Rename new Ganglia sink to 'libgangliaSink'
